Fix sync direction on initial rsync
It turns out that the order of the source and destination in the rsync command is important and changes the behavior of the syncing operation. If `/etc/letsencrypt` is empty (which it would on first launch), rsync won't fetch the files from the destination.

This commit changes the order of the first rsync command to force rsync to fetch the files.